+++ src/values.rs
@@ -1,3 +1,4 @@
+use std::fmt;
 use std::rc::Rc;
 
 use crate::parser::ExprNode;
@@ -26,5 +27,48 @@ impl From<TrackedVector> for LongleafValue {
 impl From<f64> for LongleafValue {
     fn from(f: f64) -> LongleafValue {
         LongleafValue::Float(f)
+    }
+}
+
+impl fmt::Display for LongleafValue {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
+        match self {
+            LongleafValue::Float(val) => write!(f, "{}", val),
+            LongleafValue::FunctionDefinition(args, _expr) => {
+                let len = args.names.len();
+                let plural = if len > 1 { "s" } else { "" };
+                write!(f, "(function which takes {} argument{})", len, plural)
+            }
+            LongleafValue::FloatList(vals) => {
+                write!(f, "[")?;
+                if vals.is_empty() {
+                    write!(f, "]")?;
+                    return Ok(());
+                }
+
+                if vals.len() < 10 {
+                    let mut iter = vals.iter();
+                    write!(f, "{}", iter.next().unwrap())?;
+
+                    for next in iter {
+                        write!(f, ", {}", next)?;
+                    }
+                } else {
+                    write!(f, "{}", vals[0])?;
+
+                    for v in &vals[1..5] {
+                        write!(f, ", {}", v)?;
+                    }
+
+                    write!(f, ", ...")?;
+
+                    for v in &vals[vals.len() - 5..] {
+                        write!(f, ", {}", v)?;
+                    }
+                }
+
+                write!(f, "]")
+            }
+        }
     }
 }
